+++ tests/testthat/test-pdf_add_origami.R
@@ -1,0 +1,14 @@
+test_that("`pdf_add_origami()`", {
+    on.exit(rm_temp_pdfs(), add = TRUE)
+
+    f1 <- tempfile(fileext = ".pdf")
+    grDevices::pdf(f1, width = 11, height = 8.5)
+    pnpmisc:::grid.mock_sbgj()
+    invisible(grDevices::dev.off())
+
+    f2 <- pdf_add_origami(f1)
+
+    expect_equal(qpdf::pdf_length(f2), 1L)
+    expect_equal(pdf_height(f2, numeric = TRUE), 8.5)
+    expect_equal(pdf_width(f2, numeric = TRUE), 11)
+})